requirements.txt
# Core
Django>=4.2,<5.0
Pillow>=10.0

# Production server
gunicorn>=21.0

# PostgreSQL
psycopg2-binary>=2.9

# Static files (serves directly from Django in production without a separate nginx for static)
whitenoise>=6.6

# Environment variables
python-decouple>=3.8
